+++ scrabble_cheater/frontend/constants/actions.js
@@ -1,15 +1,13 @@
-// Board
+// Board / Tile
 export const TOGGLE_GAME_TYPE = 'TOGGLE_GAME_TYPE'
-export const UPDATE_RACK = 'UPDATE_RACK'
 export const CHANGE_COORDINATE_VALUE = 'CHANGE_COORDINATE_VALUE'
 export const SET_MOVE_DIRECTION = 'SET_MOVE_DIRECTION'
-
-// Tiles
 export const MAKE_TILE_EDITABLE = 'MAKE_TILE_EDITABLE'
 
-// Rack
+// Rack / Wordlist
+export const UPDATE_RACK = 'UPDATE_RACK'
+export const HIGHLIGHT_SUGGESTED_WORD = 'HIGHLIGHT_SUGGESTED_WORD'
+
+// Websockets
 export const SENT_TABLE_DATA = 'SENT_TABLE_DATA'
 export const RECEIVED_SUGGESTED_WORDS = 'RECEIVED_SUGGESTED_WORDS'
-
-// Wordlist
-export const HIGHLIGHT_SUGGESTED_WORD = 'HIGHLIGHT_SUGGESTED_WORD'
